-Added group files Services and Models (also swift files inside them) -Linked structs and methods in Services and Models with MovieQuizViewController -Created QuestionFactoryProtocol
