+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,13 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Almira Khafizova on 11.06.23.
+//
+
+import Foundation
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
