+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Almira Khafizova on 11.06.23.
+//
+
+import Foundation
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
